mcp_server/requirements.txt
requests
httpx
fastmcp
uvicorn
PyYAML